% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_nonempty.R
\name{coalesce_nonempty}
\alias{coalesce_nonempty}
\title{Coalesce First Non-empty String}
\usage{
coalesce_nonempty(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character string that is the first non-empty, non-NA element. Returns NA if none found.
}
\description{
Returns the first element in a character vector that is neither NA nor an empty string ("").
}
\examples{
coalesce_nonempty(c(NA, "", "apple", "banana"))
coalesce_nonempty(c(NA, ""))
}
