% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bernoulli.R
\name{fit_bernoulli}
\alias{fit_bernoulli}
\title{Fit Bernoulli Distribution}
\usage{
fit_bernoulli(x)
}
\arguments{
\item{x}{A logical or binary numeric vector (0/1).}
}
\value{
A list with one element \code{p}, the estimated probability of success.
}
\description{
Estimates the success probability of a Bernoulli distribution using maximum likelihood estimation (MLE).
}
\examples{
fit_bernoulli(c(TRUE, FALSE, TRUE, TRUE))
fit_bernoulli(c(1, 0, 1, 1))
}
