% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_zero.R
\name{gt_zero}
\alias{gt_zero}
\title{Truncate Values Below Zero}
\usage{
gt_zero(x)
}
\arguments{
\item{x}{A numeric value or vector.}
}
\value{
A numeric vector with values < 0 replaced by 0.
}
\description{
Replaces any numeric values below zero with zero.
}
\examples{
gt_zero(c(-1, 0, 2, -3, 5))
}
